+++ include/seqan3/contrib/parallel/buffer_queue.hpp
@@ -21,7 +21,7 @@
 #include <vector>
 
 #include <seqan3/core/bit_manipulation.hpp>
-#include <seqan3/core/parallel/detail/spin_delay.hpp>
+#include <seqan3/utility/parallel/detail/spin_delay.hpp>
 #include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/range/container/concept.hpp>
 #include <seqan3/std/algorithm>
