+++ include/seqan3/core/parallel/detail/latch.hpp
@@ -15,7 +15,7 @@
 #include <atomic>
 #include <cassert>
 
-#include <seqan3/core/parallel/detail/spin_delay.hpp>
+#include <seqan3/utility/parallel/detail/spin_delay.hpp>
 #include <seqan3/std/new>
 
 namespace seqan3::detail
