+++ include/seqan3/core/parallel/detail/reader_writer_manager.hpp
@@ -18,7 +18,7 @@
 #include <seqan3/std/new>
 
 #include <seqan3/core/parallel/detail/latch.hpp>
-#include <seqan3/core/parallel/detail/spin_delay.hpp>
+#include <seqan3/utility/parallel/detail/spin_delay.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
 
 namespace seqan3::detail
